Move `local` module out of obsolete `shape` module
