--- crates/fj-kernel/src/shape/mod.rs
@@ -1,7 +1,0 @@
-//! The API used for creating and manipulating shapes
-//!
-//! See [`Shape`], which is the main entry point to this API.
-
-mod local;
-
-pub use self::local::LocalForm;
